[38-distance_to_PB] 1. 코드 정리 마커 클릭시 발생하는 이벤트
