+++ .idea/misc.xml
@@ -23,6 +23,7 @@
         <entry key="..\:/Users/user/AndroidStudioProjects/pozi_android/app/src/main/res/layout/activity_main.xml" value="0.25364583333333335" />
         <entry key="..\:/Users/user/AndroidStudioProjects/pozi_android/app/src/main/res/layout/item_pb_viewpager.xml" value="0.176" />
         <entry key="..\:/Users/user/AndroidStudioProjects/pozi_android/app/src/main/res/layout/item_photobooth_detail_viewpager.xml" value="0.4" />
+        <entry key="..\:/Users/user/AndroidStudioProjects/pozi_android/app/src/main/res/layout/select_map_application_bottom_sheet.xml" value="0.25364583333333335" />
       </map>
     </option>
   </component>
